# Fix/improve project allowed to scope performance