--- a/app/controllers/users_controller.rb
+++ b/app/controllers/users_controller.rb
@@ -65,10 +65,14 @@ class UsersController < ApplicationController
   end
 
   def show
-    # show projects based on current user visibility
+    # show projects based on current user visibility.
+    # But don't simply concatenate the .visible scope to the memberships
+    # as .memberships has an include and an order which for whatever reason
+    # also gets applied to the Project.allowed_to parts concatenated by a UNION
+    # and an order inside a UNION is not allowed in postgres.
     @memberships = @user.memberships
                         .where.not(project_id: nil)
-                        .visible(current_user)
+                        .where(id: Member.visible(current_user))
 
     if can_show_user?
       @events = events