--- a/app/models/members/scopes/visible.rb
+++ b/app/models/members/scopes/visible.rb
@@ -43,16 +43,16 @@ module Members::Scopes
       private
 
       def visible_for_non_admins(user)
-        view_members = Project.where(id: Project.allowed_to(user, :view_members))
-        manage_members = Project.where(id: Project.allowed_to(user, :manage_members))
+        view_members = Project.allowed_to(user, :view_members)
+        manage_members = Project.allowed_to(user, :manage_members)
 
         project_scope = view_members.or(manage_members)
 
-        Member.where(project_id: project_scope.select(:id))
+        where(project_id: project_scope.select(:id))
       end
 
       def visible_for_admins
-        Member.all
+        all
       end
     end
   end