Reworks the SQL statement underlying

`Project.allowed_to?(some_user, :some_permission)`

that showed poor performance in some datasets. 

### Background

The problem manifests in statements where the aforementioned statement takes part as a subquery in combination with a `LIMIT`, as e.g. in 

```sql

SELECT "work_packages"."id"                      AS t0_r0,
       "work_packages"."type_id"                 AS t0_r1,
       "work_packages"."project_id"              AS t0_r2,
       "work_packages"."subject"                 AS t0_r3,
       "work_packages"."description"             AS t0_r4,
       "work_packages"."due_date"                AS t0_r5,
       "work_packages"."category_id"             AS t0_r6,
       "work_packages"."status_id"               AS t0_r7,
       "work_packages"."assigned_to_id"          AS t0_r8,
       "work_packages"."priority_id"             AS t0_r9,
       "work_packages"."version_id"              AS t0_r10,
       "work_packages"."author_id"               AS t0_r11,
       "work_packages"."lock_version"            AS t0_r12,
       "work_packages"."done_ratio"              AS t0_r13,
       "work_packages"."estimated_hours"         AS t0_r14,
       "work_packages"."created_at"              AS t0_r15,
       "work_packages"."updated_at"              AS t0_r16,
       "work_packages"."start_date"              AS t0_r17,
       "work_packages"."responsible_id"          AS t0_r18,
       "work_packages"."position"                AS t0_r19,
       "work_packages"."story_points"            AS t0_r20,
       "work_packages"."remaining_hours"         AS t0_r21,
       "work_packages"."budget_id"               AS t0_r22,
       "work_packages"."derived_estimated_hours" AS t0_r23,
       "work_packages"."schedule_manually"       AS t0_r24,
       "work_packages"."parent_id"               AS t0_r25,
       "work_packages"."duration"                AS t0_r26,
       "work_packages"."ignore_non_working_days" AS t0_r27,
       "work_packages"."derived_remaining_hours" AS t0_r28,
       "projects"."id"                           AS t1_r0,
       "projects"."name"                         AS t1_r1,
       "projects"."description"                  AS t1_r2,
       "projects"."public"                       AS t1_r3,
       "projects"."parent_id"                    AS t1_r4,
       "projects"."created_at"                   AS t1_r5,
       "projects"."updated_at"                   AS t1_r6,
       "projects"."identifier"                   AS t1_r7,
       "projects"."lft"                          AS t1_r8,
       "projects"."rgt"                          AS t1_r9,
       "projects"."active"                       AS t1_r10,
       "projects"."templated"                    AS t1_r11,
       "projects"."status_code"                  AS t1_r12,
       "projects"."status_explanation"           AS t1_r13
FROM "work_packages"
         LEFT OUTER JOIN "projects" ON "projects"."id" = "work_packages"."project_id"
WHERE "work_packages"."project_id" IN (
	SELECT "projects"."id"
                                       FROM "projects"
                                                LEFT OUTER JOIN "members"
                                                                ON "projects"."id" = "members"."project_id" AND
                                                                   "members"."user_id" = 60026 AND
                                                                   "members"."entity_type" IS NULL AND
                                                                   "members"."entity_id" IS NULL AND
                                                                   "projects"."active" = TRUE
                                                INNER JOIN "enabled_modules"
                                                           ON "projects"."id" = "enabled_modules"."project_id" AND
                                                              "enabled_modules"."name" IN ('work_package_tracking') AND
                                                              "projects"."active" = TRUE
                                                INNER JOIN "role_permissions"
                                                           ON "role_permissions"."permission" IN ('view_work_packages')
                                                INNER JOIN "roles" "permission_roles"
                                                           ON "permission_roles"."id" = "role_permissions"."role_id"
                                                LEFT OUTER JOIN "member_roles" ON "members"."id" = "member_roles"."member_id"
                                                LEFT OUTER JOIN "roles" "assigned_roles"
                                                                ON "assigned_roles"."id" = "permission_roles"."id" AND
                                                                   "projects"."active" = TRUE AND
                                                                   ("assigned_roles"."id" = "member_roles"."role_id" OR (
                                                                    "projects"."public" = TRUE AND
                                                                    "assigned_roles"."builtin" = 1 AND
                                                                    "member_roles"."id" IS NULL))
                                       WHERE "assigned_roles"."id" IS NOT NULL
)
  AND (((work_packages.subject ILIKE '%6%' OR projects.name ILIKE '%6%')))
ORDER BY work_packages.updated_at, work_packages.id DESC
LIMIT 10 
```

A statement like this is issued when using the work package autocompleter inside text fields (i.e. typing `#6` in a comment field) but similar requests are issued e.g. on
* /api/v3/work_packages
* /api/v3/queries/default

As such, in the problematic datasets, central parts of the application are affected for non admin users.

The reason for the poor performance is the query planner assuming that it will be faster to go over the work packages from the top and find those that match the filtered for (`work_packages.subject ILIKE '%6%' OR projects.name ILIKE '%6%'`) and afterwards check if they are in a project the user is allowed to see. This is true even though only one project might be returned by the subselect and in the dataset in a lot of cases is. Running `ANALYZE` on the tables does not help in this case. 

This behaviour is a known shortcoming of PostgreSQL in some scenarios (see https://www.postgresql.org/message-id/flat/CA%2BU5nMLbXfUT9cWDHJ3tpxjC3bTWqizBKqTwDgzebCB5bAGCgg%40mail.gmail.com and https://stackoverflow.com/questions/60115419/postgres-slower-when-a-limit-is-set-how-to-fix-besides-adding-a-dummy-order-by/60118336#60118336)

Since PostgreSQL does not support query hints, the query planner cannot be coerced into first running the subselect.

The subselect has a shortcoming of itself, though in that it needs to have a LEFT JOIN for the members and an OR when joining the roles to support both actual project members and their roles as well as the builtin roles on public projects.

## Solution

For those two reasons, it seems prudent to rewrite the subselect. In the PR, a UNION ALL based approach is chosen to separate the two cases (actual project member roles vs builtin roles). Then then also leads to improved performance as the query planner choses a better execution plan.

The resulting query looks like this

```SQL
SELECT "work_packages"."id" AS T0_R0,
	"work_packages"."type_id" AS T0_R1,
	"work_packages"."project_id" AS T0_R2,
	"work_packages"."subject" AS T0_R3,
	"work_packages"."description" AS T0_R4,
	"work_packages"."due_date" AS T0_R5,
	"work_packages"."category_id" AS T0_R6,
	"work_packages"."status_id" AS T0_R7,
	"work_packages"."assigned_to_id" AS T0_R8,
	"work_packages"."priority_id" AS T0_R9,
	"work_packages"."version_id" AS T0_R10,
	"work_packages"."author_id" AS T0_R11,
	"work_packages"."lock_version" AS T0_R12,
	"work_packages"."done_ratio" AS T0_R13,
	"work_packages"."estimated_hours" AS T0_R14,
	"work_packages"."created_at" AS T0_R15,
	"work_packages"."updated_at" AS T0_R16,
	"work_packages"."start_date" AS T0_R17,
	"work_packages"."responsible_id" AS T0_R18,
	"work_packages"."position" AS T0_R19,
	"work_packages"."story_points" AS T0_R20,
	"work_packages"."remaining_hours" AS T0_R21,
	"work_packages"."budget_id" AS T0_R22,
	"work_packages"."derived_estimated_hours" AS T0_R23,
	"work_packages"."schedule_manually" AS T0_R24,
	"work_packages"."parent_id" AS T0_R25,
	"work_packages"."duration" AS T0_R26,
	"work_packages"."ignore_non_working_days" AS T0_R27,
	"work_packages"."derived_remaining_hours" AS T0_R28,
	"projects"."id" AS T1_R0,
	"projects"."name" AS T1_R1,
	"projects"."description" AS T1_R2,
	"projects"."public" AS T1_R3,
	"projects"."parent_id" AS T1_R4,
	"projects"."created_at" AS T1_R5,
	"projects"."updated_at" AS T1_R6,
	"projects"."identifier" AS T1_R7,
	"projects"."lft" AS T1_R8,
	"projects"."rgt" AS T1_R9,
	"projects"."active" AS T1_R10,
	"projects"."templated" AS T1_R11,
	"projects"."status_code" AS T1_R12,
	"projects"."status_explanation" AS T1_R13
FROM "work_packages"
LEFT OUTER JOIN "projects" ON "projects"."id" = "work_packages"."project_id"
WHERE "work_packages"."project_id" IN
		(SELECT "projects"."id"
			FROM "projects"
			WHERE "projects"."id" IN (
							(SELECT "projects"."id"
								FROM "members"
								INNER JOIN "projects" ON "projects"."active" = TRUE
								AND "members"."project_id" = "projects"."id"
								AND "members"."user_id" = 60026
								AND "members"."entity_type" IS NULL
								AND "members"."entity_id" IS NULL
								INNER JOIN "enabled_modules" ON "projects"."id" = "enabled_modules"."project_id"
								AND "enabled_modules"."name" IN ('work_package_tracking')
								AND "projects"."active" = TRUE
								INNER JOIN "member_roles" ON "member_roles"."member_id" = "members"."id"
								INNER JOIN "roles" ON "roles"."id" = "member_roles"."role_id"
								INNER JOIN "member_roles" "member_roles_members" ON "member_roles_members"."member_id" = "members"."id"
								INNER JOIN "role_permissions" ON "roles"."id" = "role_permissions"."role_id"
								AND (FALSE
									 OR "role_permissions"."permission" = 'view_work_packages'
									 AND "enabled_modules"."name" = 'work_package_tracking')
								
                                                                UNION ALL
 
                                                                SELECT "projects"."id"
								FROM "projects"
								INNER JOIN "enabled_modules" ON "projects"."id" = "enabled_modules"."project_id"
								AND "enabled_modules"."name" IN ('work_package_tracking')
								AND "projects"."active" = TRUE
								INNER JOIN "roles" ON "roles"."builtin" = 1
								AND "projects"."active"
								AND "projects"."public"
								AND NOT (EXISTS
									(SELECT 1
										FROM "members"
										WHERE "members"."project_id" = "projects"."id"
											AND "members"."user_id" = 60026
											AND "members"."entity_type" IS NULL
											AND "members"."entity_id" IS NULL))
								INNER JOIN "role_permissions" ON "roles"."id" = "role_permissions"."role_id"
								AND (FALSE
									 OR "role_permissions"."permission" = 'view_work_packages'
									 AND "enabled_modules"."name" = 'work_package_tracking'))))
	AND (((WORK_PACKAGES.SUBJECT ILIKE '%6%' OR PROJECTS.NAME ILIKE '%6%')))
ORDER BY WORK_PACKAGES.UPDATED_AT,
	WORK_PACKAGES.ID DESC
LIMIT 10

``` 

## Result

All results are measured on the same dataset for the same non admin user and with a hot cache.

* WorkPackage autocompleter: 
  * before: 7.5s
  * after: ~40 ms
* /api/v3/work_packages?pageSize=100 
  * before: 3.5s
  * after: 330ms
* /api/v3/queries/default?pageSize=100
  * before: 3.5s
  * after: 370ms  

## Observations

The performance for administrator users does not changes which isn't surprising as the statement for those users isn't altered. But interestingly, of the roughly 600ms spent for e.g. `/api/v3/work_packages?pageSize=100` over 300ms are spent calculating the overall number of work packages visible to the admin `SELECT COUNT(DISTINCT "work_packages"."id") FROM "work_packages"...`. While COUNT is notoriously slow, it might be beneficial to at least remove the `DISTINCT "work_packages"."id"` part in favour of a `*` but this is out of scope for this PR.